+++ src/TestSuite/TestCase.php
@@ -15,7 +15,7 @@ namespace Thumber\TestSuite;
 use Cake\Core\Configure;
 use Cake\Filesystem\Folder;
 use Cake\TestSuite\TestCase as CakeTestCase;
-use Reflection\ReflectionTrait;
+use Tools\ReflectionTrait;
 use Thumber\ThumbTrait;
 
 /**