+++ CHANGELOG.md
@@ -7,7 +7,9 @@
     `tests/bootstrap.php` file, using the `THUMBER_DRIVER` environment variable.
     By default, Imagick is used;
 * full support for Windows, added tests for Appveyor;
-* added `Thumber\TestSuite\TestCase::assertThumbUrl()` method.
+* added `Thumber\TestSuite\TestCase::assertThumbUrl()` method;
+* now it uses the `mirko-pagliai/php-tools` package. This also replaces
+    `mirko-pagliai/reflection`.
 
 ## 1.3 branch
 # 1.3.1